refactor: pluralize router filename
