+++ tournament_app_backend/app/app.py
@@ -10,7 +10,7 @@ from beanie import init_beanie
 
 from models import __document_models__
 from routers.events import router as eventsRouter
-from routers.tournament import router as tournamentsRouter
+from routers.tournaments import router as tournamentsRouter
 from routers.auth import router as AuthRouter
 
 from fastapi_pagination import add_pagination
